merge V3.4.0

+++ native/cocos/scene/SubModel.cpp
@@ -1,8 +1,8 @@
 /****************************************************************************
  Copyright (c) 2021 Xiamen Yaji Software Co., Ltd.
- 
+
  http://www.cocos.com
- 
+
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated engine source code (the "Software"), a limited,
  worldwide, royalty-free, non-assignable, revocable and non-exclusive license
@@ -10,10 +10,10 @@
  not use Cocos Creator software for developing other software or tools that's
  used for developing games. You are not granted to publish, distribute,
  sublicense, and/or sell copies of Cocos Creator.
- 
+
  The software or tools in this License Agreement are licensed, not sold.
  Xiamen Yaji Software Co., Ltd. reserves all rights not expressly granted to you.
- 
+
  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -39,16 +39,16 @@ SubModel::~SubModel() {
     delete _subMesh;
 }
 
-gfx::Shader *SubModel::getShader(int index) const {
-    if (index >= static_cast<int>(_shaders.size())) {
+gfx::Shader *SubModel::getShader(uint index) const {
+    if (index >= _shaders.size()) {
         return nullptr;
     }
 
     return _shaders[index];
 }
 
-Pass *SubModel::getPass(int index) const {
-    if (index >= static_cast<int>(_passes.size())) {
+Pass *SubModel::getPass(uint index) const {
+    if (index >= _passes.size()) {
         return nullptr;
     }
